requirements.txt
fastapi
uvicorn[standard]
aiogram
sqlalchemy
aiomysql
jinja2
pyyaml
pydantic
pydantic-settings
loguru
python-dotenv
cryptography
httpx
